Extract option reading into a helper class.
Also combine the option reading code for field options and
enum options.

This is intended to make options reusable for ProfileParser.
